Excepciones y control de ellas

+++ gestorCursosApi/src/main/java/com/pejo/anna/gestorCursosApi/servicios/ApiServiceImpl.java
@@ -2,10 +2,14 @@ package com.pejo.anna.gestorCursosApi.servicios;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.pejo.anna.gestorCursosApi.entidades.Usuario;
 import com.pejo.anna.gestorCursosApi.repositorios.UsuarioRepository;
 
+import lombok.extern.java.Log;
+
+@Log
 @Service
 public class ApiServiceImpl implements ApiService{
 	
@@ -33,6 +37,7 @@ public class ApiServiceImpl implements ApiService{
 	@Override
 	public void anadirUsuario(Usuario usuario) {
 		usuarioRepository.save(usuario);
+		log.info("Se ha añadido el usuario " + usuario);
 	}
 
 
@@ -56,7 +61,12 @@ public class ApiServiceImpl implements ApiService{
 
 
 	@Override
+	@Transactional
 	public void borrarUsuario(Long id) {
+		if(!usuarioRepository.existsById(id)) {
+			log.warning("Se ha intentado borrar un usuario que no existe: " + id);
+			throw new ServiciosException("No existe el usuario a borrar");
+		}	
 		usuarioRepository.deleteById(id);
 	}
